Add files via upload
The Mouse Coding Project involves analyzing a dataset of mouse data collected from a study on various drug regimens. The primary goals of the project are to clean and process the data, perform statistical analyses, identify potential outliers, and visualize the results through various plots. This project provides insights into the effectiveness of different treatments in reducing tumor volume in mice.